requirements.txt
flask==2.3.3
werkzeug==2.3.7
rembg==2.0.50
Pillow==9.5.0
flask-cors==4.0.0 